requirements.txt
absl-py==1.2.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.2
aspose-cells==22.8.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
binaryornot==0.4.4
bleach==4.1.0
blinker==1.4
boto3==1.21.23
botocore==1.24.23
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.14
cookiecutter==2.1.1
cookiejar==0.0.3
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==36.0.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==21.1
django-location==2.1.1
django-mailbox==4.8.2
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-social-auth==0.7.28
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dynaconf==3.1.7
entrypoints==0.4
et-xmlfile==1.1.0
facebook-sdk==3.1.0
filelock==3.6.0
flatbuffers==2.0.7
fontawesomefree==6.1.2
gast==0.4.0
Gears==0.7.2
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
googletrans-temp==2.3.0
grpcio==1.49.0
gTTS==2.2.4
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2022.9.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
huepy==1.2.1
hvac==0.11.2
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
importlib-metadata==4.12.0
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==1.0.0
joblib==1.2.0
JPype1==1.4.0
jsonfield==3.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
knox==0.1.14
libclang==14.0.6
libretranslatepy==2.1.1
loguru==0.6.0
lxml==4.8.0
Markdown==3.3.3
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib-inline==0.1.3
mechanize==0.4.8
mistune==0.8.4
MouseInfo==0.1.3
Naked==0.1.31
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.11.0
nltk==3.7
notebook==6.4.10
numpy==1.21.5
oauth2==1.9.0.post1
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pager==3.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
po-excel-translate==3.3
polib==1.1.1
prometheus-client==0.13.1
prompt-toolkit==1.0.14
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
py-translate==1.0.3
PyAccessPoint==0.2.5
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.15.0
pydeck==0.7.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.11.2
pyhostapdconf==1.0
pyicloud-dwoh==0.4.1
PyInquirer==1.0.3
PyJWT==2.3.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyqt5-tools==5.15.4.3.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-core==0.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-openid==2.2.5
python-slugify==6.1.2
python3-openid==3.2.0
pytube==12.1.0
pytube3==9.6.4
pytweening==1.0.4
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
pywifi==1.1.12
pywin32==303
pywinpty==2.0.5
PyYAML==3.13
pyzmq==22.3.0
qrcode==7.3.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
regex==2022.8.17
requests==2.25.1
requests-oauthlib==1.3.1
resolvelib==0.5.4
rfc3986==1.5.0
rsa==4.9
s3transfer==0.5.2
selenium==4.4.3
semver==2.13.0
Send2Trash==1.8.0
shellescape==3.8.1
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
streamlit==1.7.0
streamlit-aggrid==0.2.2.post2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0rc3
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
tflearn==0.5.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
translate==3.6.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.2
uuid==1.30
validators==0.18.2
virtualenv==20.13.4
watchdog==2.1.6
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==5.2.0
widgetsnbextension==3.6.0
win32-setctime==1.1.0
wireless==0.3.3
WMI==1.5.1
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
